[ownership] ref_to_* and *_to_ref produce values with OwnershipKind::Unowned so should be treated as UnownedInstantaneousUse.
The reason why these values produce a value with unowned ownership is that they
perform a conversion without taking ownership of the underlying
value. Semantically we do this to ensure that before anyone tries to use one of
these in an owned or gauranteed position they will need to perform some sort of
transform on the value semantically showing that it has ownership of the
value via a copy or some other operation.