Revert "[sil-combine] Update switch_enum_addr -> switch_enum for loadable types for ownership."
This reverts commit 0a0f8cffc1e31c7d5876d073a7d1cd75608759f0.

Currently the memory lifetime verifier (I believe due to the switch_enum) is not
caring that we are not emitting destroy_addr for an enum address along paths
where we know the underlying enum case is trivial.

This transform eliminated that switch_enum_addr causing the memory lifetime
verifier to then trigger.

Disabling to unblock the bots!