ClangImporter: unify language standard selection (NFCI)
This simply extracts the shared portion of the language version setup.
The state of the ObjC interop determines if we are in ObjectiveC or C
mode or ObjectiveC++ or C++ mode depending on C++ interop.  However, if
Objective-C interop is enabled, we must enable ARC and set the ObjC
runtime version according to the target.  This simply shares more of the
setup and is meant as a simplification for pairing the language version.