[ConstraintSystem] Increase impact of implicit Double<->CGFloat associated with contextual type
Consider conversion with context to be worse by default
because it means that expression as a whole couldn't get
to the expected type.

This also helps to disambiguate situations with
multiple solutions where one of them has a contextual
mismatch that requires implicit conversion and
another produces correct contextual type but requires
implicit conversion inside to get there.