[Type checker] Ensure we sort semantic members properly.
When evaluating the semantic members request, we would sort the members
before we return them to ensure a stable ordering for (e.g.) vtable
layout. However, when not in a source file, we would assume that the
resulting members were always sorted. However, they are not always
sorted, because we write the iterative decl contexts directly rather
than writing the semantic members.

Fixes rdar://72833032.