[ome] Explicitly set to OwnershipKind::None the ownership forwarded by all instructions that inherit from OwnershipForwardingMixin.
I thought it would be enough to change SILValue::getOwnershipKind() to return
None when we are not in OSSA, but I found that is not sufficient. In some cases,
users will call getOwnershipKind() directly on these classes and will get the
stored value (which has not been changed) instead of the correct ownership,
OwnershipKind::None.

I ran into this when one of the RAUW utilities when done a path that wanted
Owned ownership when we were in a non-ossa function and then (happily I may add)
hit an assert.