[cmake] This ld64 LTO flag is only used on Darwin.
If a SDK has no specified variant link flags and LTO is specified, then
we attempt to pass Apple platform specific linker flags (in this case,
`object_path_lto`) to the platform linker, which may indeed not be ld64.

SWIFT_USE_LINKER is set empty on Apple platforms, so it is somewhat
onerous to exclude every possible linker here. We could affirmatively set
cases to the branch, but then the next platform to be supported may run
into this platform.

The simplest thing to do is just to create a specific case for Apple
platforms, and the reasonable default is to assume no specific host
linker flags are required.