[libSyntax] Add an unowned *Ref version of AbsoluteRawSyntax and SyntaxData
AbsoluteRawSyntaxRef doesn't keep the underlying RawSyntax node alive,
avoiding some of the expensive reference counting. The user of these
types is responsible for ensuring that the RawSyntax nodes stay in
memory, e.g. by keeping the SyntaxArena in which they live alive.

SyntaxDataRef also doesn't keep its parent alive and requires the caller
to make sure that the parent always outlives a child. This property is
nearly always implicitly given when walking a syntax tree.