[Stats] Fix a second/nanoseconds bug in process-stats-dir
Times are expected to be represented in nanoseconds in process-stats-dir,
but times with the 'swift.time.' prefix (e.g. times for specific
requests) were not converted.

It appears the reasons that this hasn’t been caught so far, is that
these times are not shown in Swift-CI's please test compiler performance
report.