ABI: convert some `unsigned long` to `unsigned long long`
This path is not yet ABI stable so take advantage to correct the
storage.  This ABI compatible on most platforms, however, is not
compatible on Windows, which is a ILLP64 platform rather than ILP64.
This extends the storage to 64-bits from 32-bits.  In the process, it
avoids some warnings on the windows builds as well.